parlib: Make uth mutex/CV initializers C++ safe (XCC)
GCC C++ can't handle { { .field = value } }.  It can handle one {}, but not
two.  C doesn't have this problem.  The issue comes up when someone does
something like:

	struct my_struct {
		uth_mutex_t foo;
	}
	#define MY_INITIALIZER {UTH_MUTEX_INIT}

icu does this with their umutex.h implementation.  There it is for
pthreads.

For us, we can get away with the non-specific initializers, so long as we
put the parlib_once_t first.

Rebuild your toolchain.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>