pthreads: Remove sched_policy and sched_priority
We never actually supported this - we just faked it.  We'll continue to
fake it, but won't track the faked values in the thread struct.

If it turns out that the pthreads interface is the best way for apps to
influence a 2LS, then we can hook these functions up to real 2LS ops.
Until then, we'll go with the "application-specific" nature of the 2LS,
where the 2LS itself is made to work with a particular app.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>