net: tcp: Remove the Syn_received state
I think Presotto's note was right: you should send a reset.  The situation
is when we sent a SYN, then we receive a SYN without an ACK.  That was the
only time we'd set Syn_received, and a reset is safer.  I think.

It also removes a bunch of code that pretty much never runs, except during
exceptions/attacks.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>