net: tcp: Don't scale SRTT and MDEV
The old code was keeping the values of SRTT and MDEV (really RTTVAR)
scaled, as if they were multiplied by their alpha and beta (bravo) factors.
This will be a minor issue when we try to use TS for RTTM.

Also, we weren't scaling those values back, so if you looked at e.g.
/net/tcp/1/status, you'd get a scaled value for the RTT.

Also, the RTO calculation looked off slightly.  RFC 6298.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>