net: tcp: Use timestamps for the RTTM
This greatly increases the accuracy of our RTTM.  Our RTT granularity for
the non-TS case was 50 msec.  So for machines less than 50 msec away,
you're getting 50 as your srtt.

The one thing I'm dubious about is whether the expected_samples business is
worth the cost of the divides.  RFC 7323 said we should do something, not
must.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>